requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.21.1
coverage==5.2
coveralls==2.1.1
Django==3.0.8
djangorestframework==3.11.0
docopt==0.6.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
